Pre-declaration of the classes, FriendList and PostTree, and the converion of the by-values to pointers.
The classes FriendList and PostTree were not declared befor the struct UserNode. This caused an error because due to sequential compilation, as the compiler reache the friends and posts by-value parameters of the UserNode, it didn't have any idea what they were, because the classes of these two instances were declared later previously. So I predeclared them to let the compiler know that such data type exists but will be defined later. I also changed the by-values to pointers because at the point the compiler reaches the parameters, it has no idea what the size of these by-value parameters would be. So I changed them into pointers, so that the compiler doesn't focus on the size.