Prepared the main function.
Created a user list of the class UserList, which basically is a linked list.
The choice is set -1 initially.
Created  a while loop for the user to enter his choice on each turn.
Also made a precautionary measure, if a string is entered instead of integer, then a warning will be given and the current on-going iteration would be skipped to the other iteration.