Filled the first two choices
The first choice asks the user to enter the name of the user to be added, gives a warning if the username is typed empty and skips the iteration in the case the name entered is actually empty. In the case the name is entered correctly it adds it to the linked list of the users.
The second choice is used to display all the users that were added to the UserList linked list. It is used to travese through the whole list and give the relevant data regarding the user that is traversesd.