Ensuring the FriendLIst and PostTree are already initialized.
There might be cases where the FriendList friend does not exist and there fore it becomes impossible for to access the friends list. Similarly the same goes for the PostTree, there might be cases where the PostTree might not exist already and it would be impossible to add a post, in the AVL Tree when there is not PostTree. For that we have created the ouline functions, ensure_friends_initialized() and ensure_posts_initialized(), to ensure if the list or the tree exist. In the case if they don't, we call the outline functioons to create them and then perform the required operation. I have made them outline and defined them after the class. Because the function includes the use of the class pointers which are not defined before the classes in the UserNode structure.