Some changes to differentiate the CEO and the simple user of the social networking system
I added string parameter password to the UserNode structure. Each user is assigned his own account password so that it could be easier for him to enter the social networking system.
There is an authenticate function that is added to ensure if the password of a user with a particular id is correct or not. This authenticate function is used in the login_user function where the id and password is asked from the user which is then checked if it is correct or not using the authenticate function.
Also made some changes in the main function to set it in accordance with the choices of the ceo or user.
CEO gets the options to add a user, view the list of users and other three CEO options.
Where as the simple user is allowed to deal with his system in the restricted limits.